+++ packages/RSuite/NEWS.md
@@ -2,6 +2,7 @@
   * Unregistering repo adapters functionality added.
   * Dependency resolving when using multiple repos. Previously if a subdependency
     was present in a preceding repository R Suite was unable to find it.
+  * rsuite_get_os_info added to retrieve OS release/version etc.
 
 # RSuite 0.31 (2018-07-26)
   * Handling of old (not available in currently selected repositories) packages